pull more information from sessionize
