+++ .github/workflows/backend-test.yml
@@ -25,4 +25,4 @@ jobs:
       uses: gradle/gradle-build-action@v2
 
     - name: Build
-      run: ./gradlew :backend:service-graphql:build
+      run: ./gradlew build
