--- .github/workflows/verify.yml
@@ -1,18 +1,0 @@
-name: Github CI
-
-on: pull_request
-
-jobs:
-  verify:
-    runs-on: ubuntu-22.04
-    steps:
-    - uses: actions/checkout@v3
-      with:
-        lfs: 'true'
-
-    - uses: gradle/wrapper-validation-action@v1
-
-    - name: Validate Renovate
-      uses: rinchsan/renovate-config-validator@v0.0.12
-      with:
-        pattern: 'renovate.json'
