--- backend/service-graphql/README.md
@@ -1,14 +1,0 @@
-# Confetti's backend 🎊
-
-From the root project:
-
-**Run locally:**
-```
-./gradlew :backend:service-graphql:bootRun
-```
-
-**Deploy to Google Cloud:**
-```
-./gradlew :backend:service-graphql:appengineDeploy
-```
-
