--- backend/service-graphql/src/main/kotlin/dev/johnoreilly/confetti/backend/UpdateSchema.kt
@@ -1,9 +1,0 @@
-package dev.johnoreilly.confetti.backend
-
-import java.io.File
-
-fun main(args: Array<String>) {
-  val schema = buildSchema()
-  File("../../shared/src/commonMain/graphql/schema.graphqls").writeText(schema.print())
-}
-
