--- backend/service-graphql/src/main/kotlin/dev/johnoreilly/confetti/backend/main.kt
@@ -1,6 +1,0 @@
-package dev.johnoreilly.confetti.backend
-
-// Do not remove, this is used by bootRun
-fun main(@Suppress("UNUSED_PARAMETER") args: Array<String>) {
-    runServer()
-}
