--- backend/service-graphql/src/main/resources/application.yml
@@ -1,5 +1,0 @@
-graphql:
-  packages: "dev.johnoreilly.confetti.backend"
-
-  playground:
-    enabled: true
