+++ data/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("org.jetbrains.kotlin.jvm")
+    id("org.jetbrains.kotlin.plugin.spring")
+    id("org.jetbrains.kotlin.plugin.serialization")
+}
+
+configureCompilerOptions(17)
+
+dependencies {
+    implementation(libs.kotlinx.serialization)
+    implementation(libs.okhttp)
+    api(project(":sessionize"))
+}
