+++ service-graphql/README.md
@@ -1,0 +1,14 @@
+# Confetti's backend 🎊
+
+From the root project:
+
+**Run locally:**
+```
+./gradlew :service-graphql:bootRun
+```
+
+**Deploy to Google Cloud:**
+```
+./gradlew :service-graphql:appengineDeploy
+```
+
