+++ service-graphql/src/main/kotlin/am/backend/UpdateSchema.kt
@@ -1,0 +1,9 @@
+package am.backend
+
+import java.io.File
+
+fun main(args: Array<String>) {
+  val schema = buildSchema()
+  File("../../shared/src/commonMain/graphql/schema.graphqls").writeText(schema.print())
+}
+
