+++ service-graphql/src/main/kotlin/am/backend/firebase.kt
@@ -1,0 +1,31 @@
+package am.backend
+
+import com.google.firebase.FirebaseApp
+import com.google.firebase.FirebaseOptions
+import com.google.firebase.auth.FirebaseAuth
+import am.backend.graphql.googleCredentials
+
+
+private val lock = Object()
+private var _isInitialized = false
+
+fun String.firebaseUid(): String? {
+    if (this == "testToken") {
+        return "testUser"
+    }
+
+    synchronized(lock) {
+        if (!_isInitialized) {
+            val options = FirebaseOptions.builder().setCredentials(googleCredentials("firebase_service_account_key.json")).build()
+            FirebaseApp.initializeApp(options)
+            _isInitialized = true
+        }
+    }
+
+    return try {
+        FirebaseAuth.getInstance().verifyIdToken(this).uid
+    } catch (e: Exception) {
+        e.printStackTrace()
+        throw e
+    }
+}
