+++ service-graphql/src/main/kotlin/am/backend/main.kt
@@ -1,0 +1,6 @@
+package am.backend
+
+// Do not remove, this is used by bootRun
+fun main(@Suppress("UNUSED_PARAMETER") args: Array<String>) {
+    runServer()
+}
