+++ service-graphql/src/main/resources/application.yml
@@ -1,0 +1,5 @@
+graphql:
+  packages: "am.backend"
+
+  playground:
+    enabled: true
