+++ sessionize/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id("org.jetbrains.kotlin.jvm")
+    id("org.jetbrains.kotlin.plugin.spring")
+    id("org.jetbrains.kotlin.plugin.serialization")
+}
+
+configureCompilerOptions(17)
+
+
+dependencies {
+    implementation(libs.kotlinx.serialization)
+    implementation(libs.okhttp)
+}
+
