+++ sync/src/main/kotlin/sync/json.kt
@@ -1,0 +1,12 @@
+package sync
+
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.json.Json
+
+@OptIn(ExperimentalSerializationApi::class)
+val json = Json {
+    ignoreUnknownKeys = true
+    prettyPrint = true
+    prettyPrintIndent = "  "
+    encodeDefaults = true
+}
